extract signature matcher class to top level class
